==> notebooks/Step2_Cleaning.ipynb <==
## Cleaning Dataset

import pandas as pd
import numpy as np
import re
import string

import pickle

# nltk.data.path.append('/Users/jhonsen/Documents/DS/nltk_data/')
# nltk.download('wordnet', download_dir='/Users/jhonsen/Documents/DS/nltk_data/')

### Importing Pickled Dataframe

# Unpickle the dataframe
with open('../data/dfraw.pkl','rb') as fin:
    df = pickle.load(fin)

df.head()

print('original dimension: ',df.shape)

#### Pre-processing entries in `summary` and `title` columns in dataframe
- [x] Remove articles containing "Lists_of_" in the `title`
- [x] Remove first 24 entries, unrelated to scientists
- [x] Drop any duplicates
- [ ] Remove additional textual inconsistencies

# Remove entries with "Lists", which are not articles about scientists (person)  
df = df[~df['title'].str.contains('List')]

# Remove first 24 entries, not related to scientists
df = df[df.index>24]

# Drop duplicated rows
print(df.shape)
df.drop_duplicates(inplace=True)
print(df.shape)

# Replace \n with  space ' '
df['summary']= df['summary'].apply(lambda s: s.replace('\n',' '))

# For now, remove observations containing '== Background ==' in `summary`
df = df[~df['summary'].str.contains('==')]

# Check for articles with empty `summary` 
df[df.summary=='']

# Removing those empty summaries
df.drop(index=13119, inplace=True)
df.drop(index=13873, inplace=True)
df.drop(index=14137, inplace=True)
df.drop(index=14944, inplace=True)

# Remove observations that are not scientists, titles such as Aristotle's works, Albert Einstein's Theory of Relativity

#

# Remove observations that are entitled 'Theory_', which are not persons (scientists)

#

print('dimension after processing: ', df.shape)

# Write processed data as csv, just in case
df.to_csv('../data/df_processed.csv',index=False)

# Pickle the processed file
with open('../data/fclean.pkl','wb') as fout:
    pickle.dump(df, fout)

Next Step:
- Topic Modeling of these articles, Step3_Modeling.ipynb

---
