==> notebooks/Step3_TopicModeling.ipynb <==
## Topic Modeling with Non-negative Matrix Factorization

import pandas as pd
import numpy as np
import re
import os
import string
import pickle
from miscScripts import *

import nltk
from nltk.corpus import wordnet 
from nltk import wordpunct_tokenize

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.decomposition import TruncatedSVD
from sklearn.decomposition import NMF
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import normalize
from sklearn.decomposition import PCA
from sklearn.metrics import accuracy_score, confusion_matrix
from sklearn import svm

from matplotlib import pyplot as plt
%matplotlib inline

# nltk.data.path.append('/Users/jhonsen/Documents/DS/nltk_data/')
# nltk.download('wordnet', download_dir='/Users/jhonsen/Documents/DS/nltk_data/')

### Importing Pickled Dataframe

# Unpickle the clean data
with open('../data/fclean.pkl','rb') as fin:
    df = pickle.load(fin)

---

#### Vectorizing and Creating Document-To-Term Matrix 

# Turning texts into a list
sentences_tokens = df['summary'].tolist()

Using **TF-IDF** vectorizer
- English stop word
- consider uni-, bi-, trigrams

# TFIDF
vectorizer = TfidfVectorizer(token_pattern = r'\b[a-zA-Z]{3,}\b',
                             stop_words='english', 
                             strip_accents = 'unicode',
                             lowercase=True,
                             ngram_range=(1,3))

# Fit transform the documents
vectorizer.fit(sentences_tokens)
doc_word = vectorizer.transform(sentences_tokens)

# Save vectorizer and documents as pickle files for visualization (another notebook)
with open('../data/doc_words.pkl','wb') as fout:
    pickle.dump(doc_word, fout)

with open('../data/vectorizer.pkl', 'wb') as fout:
    pickle.dump(vectorizer, fout)

# Create Doc-Term-Matrix
dtm = pd.DataFrame(doc_word.toarray(),
             index= df.title,
            columns = vectorizer.get_feature_names())
dtm.head()

### Try TOPIC MODELING using NMF 
- at the current state 9 topics seem to make sense
- Truncated Singular Value Decomposition of this dtm matrix indicates that 9 components capture 80% of variance explained ratio (another notebook)

# Build an NMF with 9 topics

nmf = NMF(9)
nmf.fit(doc_word)
doc_topic = nmf.transform(doc_word)

# Save the model as a pickled file for visualization later (another notebook) 
with open('../data/nmf.pkl','wb') as fout:
    pickle.dump(nmf, fout)
with open('../data/doc_nmf.pkl','wb') as fout:
    pickle.dump(doc_topic, fout)

# Check components and feature vectors 
df_topic_word = pd.DataFrame(nmf.components_.round(3),
                         index=['comp'+str(k+1) for k in range(9)],
                         columns= vectorizer.get_feature_names())
df_topic_word

# Create a dataframe of scientists (observations) and TOPICS

df_sci_topic = pd.DataFrame(doc_topic.round(3),
                index= df.title,
                columns = ['comp'+str(k+1) for k in range(9)])

df_sci_topic.sample(5)

#### Convert NMF result into percentages, for intuitive inspection

# Turn these into percentages
df_Hp = getPercentages(df_sci_topic)

df_Hp.head()

# Collect the 9 topics into a col 
# this object can be used to create wordclouds
dict_topics_9 = collect_topics(nmf, vectorizer.get_feature_names(), no_top_words=10)

# Display words in NMF TOPICS 
display_topics(nmf, vectorizer.get_feature_names(), no_top_words=10)

  
Breakdown of Topics
- Topic 1 :  Scientist in academia
- Topic 2 :  ---> Comic-books scientists  
- Topic 3 :  Indian scientist
- Topic 4 :  ---> fictional scientist (Star Trek)
- Topic 5 :  European? 
- Topic 6 :  Russian
- Topic 7 :  Computer scientist
- Topic 8 :  ---->  TV
- Topic 9 :  Physicist

Note: 
- Topic-2, -4, and -8 seem to not be real scientists (persons). These were unintentionally collected during acquisition , as they must have some relations (on some level) with articles categorized under `Category:scientists`  
- Topic 5 seems to have months (april, february, etc.), which may have been people's date of births. Topic 4 
- Exploratory to double check these entries
- Cleaning the data further may fix this issue
---

#### Assign Names to Topics

# Rename Column names
df_Hp.columns = ['academia','comic','indian','fictional','european','russian','compsci','TV','physicist']

df_Hp.head()

# Turn the Name-index into a column and then reset index
df_Hp['title'] = df_Hp.index

print("Are the dimensions of the dataframes equal?")
df_Hp.shape[0] == df.shape[0]

# Resetting the indexes in dataframes
df.reset_index(inplace=True)
df.drop(columns='index', inplace=True)
df_Hp.index=np.arange(df_Hp.shape[0])

# This is the original dataframe containing unprocessed summary
df.head()

# This dataframe is the one to use for cosine similarity (below)
df_Hp.head()

# merged ataframe
dfc = pd.merge(df_Hp, df, how='left', on='title')
dfc.sample()

# Save clean dataframe
with open('../data/df_Hp.pkl','wb') as fout:
    pickle.dump(df_Hp, fout)    
with open('../data/df_combined.pkl','wb') as fout:
    pickle.dump(dfc, fout)

---

#### Find a representative example for each category

  
**Keywords** in TOPICS:  
  
Topic  1 (**Academia**)
university, professor, research, science, born, institute, department, director, college, engineering

Topic  2 (**Comica**)
marvel, comics, comic, comic books, comic books published, books published, american comic books, appearing american, appearing american comic, american comic

Topic  3 (**Indian**)
indian, science, india, academy, sciences, indian science, academy sciences, research, scientific, government india

Topic  4 (**Fictional**)
characters, star trek, trek, star, fictional, created star trek, created star, created, major characters, lists characters

Topic  5 (**European**)
german, known, french, work, march, april, august, february, december, january

Topic  6 (**Russian**)
russian, academy, sciences, chinese, academy sciences, member, soviet, academician, national, national academy

Topic 7 (**CompSci**)
computer, computer science, science, computer scientist, scientist, computing, professor computer, programming, professor computer science, american computer

Topic  8 (**TV**)
series, characters, television, television series, character, fictional, characters television series, characters television, fictional characters television, main recurring

Topic  9 (**Physicists**)
physics, theory, quantum, physicist, theoretical, prize, nobel, nuclear, einstein, nobel prize


# Example of TOPIC-1, 'Academia' 
df[df.index==1157].summary.values[0]

# keywords: university, professor, research

# Example of TOPIC-5 'European' 
df[df.index==1787].summary.values[0]

# keywords: August, September

# Example of TOPIC-6 'CompSci' 
df[df.index==5283].summary.values[0]

# keywords: computing, computer

# Example of TOPIC-3 'Indian'
df[df.index==851].summary.values[0]

# keywords: Indian, Science, scientific

# Example of TOPIC-2 'comic'
df[df.index==994].summary.values[0]

# keywords: comic, comic book, marvel

# Example of TOPIC-4 'fictional'
df[df.index==2522].summary.values[0]

# keywords: Star Trek, fiction

# Example of TOPIC-7 'TV'
df[df.index==126].summary.values[0]

# keywords: television, 

# Example of TOPIC-8 'physicist' 
df[df.index==19].summary.values[0]

# keywords: physicist, nobel prize

# Example of TOPIC-9 'Russian'
df[df.index==448].summary.values[0]

# keywords: russion, soviet, academy sciences

---

### QUERY
- First, let's see if we could locate people in the dataframe, based on cosine similarity
- Then, we check a random query outside of the dataframe

**Query 1** Hadley Wickham (is one of the observation in the dataframe)

query1= "Hadley Wickham is a statistician from New Zealand who is currently Chief Scientist at RStudio and an adjunct Professor of statistics at \
the University of Auckland,[3] Stanford University, \
and Rice University. He is best known for his development of open-source statistical analysis software packages for R (programming language) \
that implement logics of data visualisation and data transformation. Wickham's packages and writing are known for advocating a tidy data approach to data \
import, analysis and modelling methods."

df_Hp1,df_Qp1 = Recommender(df_Hp, query1, 'Hadley Wickham', vectorizer, nmf, top_n=10)

# Check the transformed query1
df_Qp1

# Compare query1 above with its top 10 similar people 
df_Hp1

# Other people similar to Hadley
df[df.index==7648].summary.values

- Hadley Wickham found himself on Wikipedia

---

**Query #2**: JD

# JD's LinkedIn Summary
query2 = "Research Chemist transitioning into data science. I have an academic background in spectroscopy and molecular modeling, \
which are the science of extracting signals out of the noise their interactions.\
I love integrating experimental techniques with computational approaches, to find actionable insights and most consistent answer to research questions. Hands-on\
laboratory work is fun, but in silico computational projects have always peaked my interest. The latter has led me to this exciting field of data science.\
I'm currently taking a deeper dive into machine learning and AI for industrial applications, especially, in the biomedical and healthcare fields.\
I'm also passionate about science communication, which is a way of “storytelling using data”. I find it enjoyable to decompose technical concepts, \
and convey them to a mixed audience, e.g., scientists in different disciplines or others without scientific backgrounds."


# Get recommendations of articles to look up
df_Hp2,df_Qp2 = Recommender(df_Hp, query2,'JD', vectorizer, nmf, top_n=10)

# Check the transformed query3
df_Qp2

# List of top 10 most scientists, similar to JD
df_Hp2

# Other people similar to JD
df[df.index==11262].summary.values

# Other people similar to JD
df[df.index==10676].summary.values

# Other people similar to JD
df[df.index==10306].summary.values

- JD's summary sounds similar to other computer scientists, w/ cosine similarity of ~0.9

---
